Harden config workflow and spec boundaries
