+++ apps/desktop/src/main/custom-mcp-store.ts
@@ -88,6 +88,27 @@ function writeManagedMcpMetadata(
   writeJsonFile(path, next as JsonRecord)
 }
 
+function repairStaleManagedMcpMetadata(
+  scope: NativeConfigScope,
+  directory: string | null | undefined,
+  metadata: Record<string, ManagedMcpMetadata>,
+  configuredNames: Set<string>,
+) {
+  const staleNames = Object.keys(metadata).filter((name) => !configuredNames.has(name))
+  if (staleNames.length === 0) return
+
+  log('warn', `Custom MCP metadata had stale entries without config: ${staleNames.join(', ')}. Repairing sidecar metadata.`)
+  try {
+    writeManagedMcpMetadata(scope, directory, (current) => {
+      const next = { ...current }
+      for (const name of staleNames) delete next[name]
+      return next
+    })
+  } catch (error) {
+    log('error', `Custom MCP metadata repair failed: ${error instanceof Error ? error.message : String(error)}`)
+  }
+}
+
 function serializeCustomMcp(mcp: CustomMcpConfig): JsonRecord {
   if (mcp.type === 'stdio') {
     if (!mcp.command?.trim()) {
@@ -163,11 +184,27 @@ export function readScopedMcps(scope: NativeConfigScope, directory?: string | nu
   const path = configPathForTarget(scope, directory)
   const config = readJsoncFile<JsonRecord>(path)
   const mcp = config.mcp
-  if (!mcp || typeof mcp !== 'object' || Array.isArray(mcp)) return []
   const metadata = readManagedMcpMetadata(scope, directory)
-  return Object.entries(mcp)
+  if (!mcp || typeof mcp !== 'object' || Array.isArray(mcp)) {
+    repairStaleManagedMcpMetadata(scope, directory, metadata, new Set())
+    return []
+  }
+  const entries = Object.entries(mcp)
+  repairStaleManagedMcpMetadata(scope, directory, metadata, new Set(entries.map(([name]) => name)))
+  return entries
     .map(([name, value]) => parseCustomMcpEntry(name, value, scope, metadata[name] || {}, directory))
     .filter((entry): entry is CustomMcpConfig => Boolean(entry))
+}
+
+function readScopedMcpConfig(
+  scope: NativeConfigScope,
+  directory: string | null | undefined,
+): Record<string, unknown> {
+  const path = configPathForTarget(scope, directory)
+  const config = readJsoncFile<JsonRecord>(path)
+  return config.mcp && typeof config.mcp === 'object' && !Array.isArray(config.mcp)
+    ? { ...(config.mcp as Record<string, unknown>) }
+    : {}
 }
 
 function updateScopedMcpConfig(
@@ -191,6 +228,26 @@ function updateScopedMcpConfig(
   writeTopLevelObjectPropertyFile(path, 'mcp', Object.keys(nextMcp).length === 0 ? null : nextMcp)
 }
 
+function restoreScopedMcpConfigAfterMetadataFailure(
+  scope: NativeConfigScope,
+  directory: string | null | undefined,
+  previousMcp: Record<string, unknown>,
+  error: unknown,
+) {
+  try {
+    updateScopedMcpConfig(scope, directory, () => ({ ...previousMcp }))
+  } catch (rollbackError) {
+    throw new Error(
+      `Custom MCP metadata write failed (${error instanceof Error ? error.message : String(error)}), and OpenCode MCP config rollback failed: ${rollbackError instanceof Error ? rollbackError.message : String(rollbackError)}`,
+      { cause: rollbackError },
+    )
+  }
+  throw new Error(
+    `Custom MCP metadata write failed after OpenCode MCP config update; restored previous MCP config: ${error instanceof Error ? error.message : String(error)}`,
+    { cause: error },
+  )
+}
+
 export function listCustomMcps(context?: RuntimeContextOptions) {
   const projectDirectory = resolveProjectDirectory(context?.directory)
   const entries = [
@@ -202,43 +259,53 @@ export function listCustomMcps(context?: RuntimeContextOptions) {
 
 export function saveCustomMcp(mcp: CustomMcpConfig) {
   assertCustomMcpContentLimits(mcp)
+  const previousMcp = readScopedMcpConfig(mcp.scope, mcp.directory)
   updateScopedMcpConfig(mcp.scope, mcp.directory, (current) => ({
     ...current,
     [mcp.name]: serializeCustomMcp(mcp),
   }))
-  writeManagedMcpMetadata(mcp.scope, mcp.directory, (current) => {
-    const next = { ...current }
-    const label = mcp.label?.trim() || undefined
-    const description = mcp.description?.trim() || undefined
-    const metadata: ManagedMcpMetadata = {
-      label,
-      description,
-      googleAuth: mcp.googleAuth === true ? true : undefined,
-      allowPrivateNetwork: mcp.allowPrivateNetwork === true ? true : undefined,
-      permissionMode: mcp.permissionMode === 'allow' ? 'allow' : undefined,
-      traceLabel: mcp.traceLabel?.trim() || undefined,
-      tracePluralLabel: mcp.tracePluralLabel?.trim() || undefined,
-    }
-    if (!metadata.label && !metadata.description && !metadata.googleAuth && !metadata.allowPrivateNetwork && !metadata.permissionMode && !metadata.traceLabel && !metadata.tracePluralLabel) {
-      delete next[mcp.name]
+  try {
+    writeManagedMcpMetadata(mcp.scope, mcp.directory, (current) => {
+      const next = { ...current }
+      const label = mcp.label?.trim() || undefined
+      const description = mcp.description?.trim() || undefined
+      const metadata: ManagedMcpMetadata = {
+        label,
+        description,
+        googleAuth: mcp.googleAuth === true ? true : undefined,
+        allowPrivateNetwork: mcp.allowPrivateNetwork === true ? true : undefined,
+        permissionMode: mcp.permissionMode === 'allow' ? 'allow' : undefined,
+        traceLabel: mcp.traceLabel?.trim() || undefined,
+        tracePluralLabel: mcp.tracePluralLabel?.trim() || undefined,
+      }
+      if (!metadata.label && !metadata.description && !metadata.googleAuth && !metadata.allowPrivateNetwork && !metadata.permissionMode && !metadata.traceLabel && !metadata.tracePluralLabel) {
+        delete next[mcp.name]
+        return next
+      }
+      next[mcp.name] = metadata
       return next
-    }
-    next[mcp.name] = metadata
-    return next
-  })
+    })
+  } catch (error) {
+    restoreScopedMcpConfigAfterMetadataFailure(mcp.scope, mcp.directory, previousMcp, error)
+  }
   return true
 }
 
 export function removeCustomMcp(target: ScopedArtifactRef) {
+  const previousMcp = readScopedMcpConfig(target.scope, target.directory)
   updateScopedMcpConfig(target.scope, target.directory, (current) => {
     const next = { ...current }
     delete next[target.name]
     return next
   })
-  writeManagedMcpMetadata(target.scope, target.directory, (current) => {
-    const next = { ...current }
-    delete next[target.name]
-    return next
-  })
+  try {
+    writeManagedMcpMetadata(target.scope, target.directory, (current) => {
+      const next = { ...current }
+      delete next[target.name]
+      return next
+    })
+  } catch (error) {
+    restoreScopedMcpConfigAfterMetadataFailure(target.scope, target.directory, previousMcp, error)
+  }
   return true
 }
